프로그래머스12954: x만큼 간격이 있는 n개의 숫자

+++ Programmers/PGM12954.java
@@ -8,7 +8,7 @@ public class PGM12954 {
     public static long[] solution(int x, int n) {
         long[] answer = new long[n];
         for (int i = 0; i <n ; i++) {
-            answer[i]=x+x*i;
+            answer[i]=x+ (long) x *i;
         }
         return answer;
     }
